+++ MusicPlayer/Library/SongDTO.cs
@@ -6,6 +6,6 @@ namespace MusicPlayer.Library
     public class SongDTO
     {
         public string Title { get; set; }
-        public int SongID { get; set; }
+        public long SongID { get; set; }
     }
 }
